pyproject.toml
[project]
name = "mcp-semantic-segmentation"
version = "0.1.0"
description = "MCP server for semantic segmentation"
authors = [
    {name = "Seiya Kumada", email = "seiya.kumada@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp>=1.0.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "opencv-python>=4.8.0",
    "pillow>=10.0.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ['py313']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
